Move link type to Link file
