+++ src/components/material-kit/Navbar/common/BrandLink.tsx
@@ -33,11 +33,11 @@ import React from 'react';
 
 import { Link as GatsbyLink } from 'gatsby-link';
 
-import { Link } from '../../../../assets/oldTheme';
+import { LinkPlop } from '../../../';
 
 import { MKBox, MKTypography } from '../..';
 
-export type BrandContent = Required<Omit<Link, 'description'>> & {
+export type BrandContent = Required<Omit<LinkPlop, 'description'>> & {
   logo: { primary: string; light: string };
 };
